Update s3 Lambda Recorder
